+++ changelog.md
@@ -2,6 +2,7 @@
 
 # v0.12.2
 - Fixed validate parent entity owner id in sync queue actions.
+- Added context info in bad request exception by entity max count exceeded.
 
 # v0.12.1
 - Added validate to delete actions in restriction to queue actions.